event planning

+++ src/main/java/Gella/Tailor_assistant/View/NewOrderWindow.java
@@ -465,8 +465,9 @@ public class NewOrderWindow extends JFrame{
 			@Override
 			public void actionPerformed(ActionEvent e) {
 			try {	
-			double duration=Float.parseFloat(execTimeField.getText())*3600000;	
+			long duration=Long.parseLong(execTimeField.getText())*3600000;
 			dates= calendarController.getFreeDates(duration);
+		//	for(Event d:dates) System.out.println(d.getStart().toString()+"  " + new Date (d.getStart().getTime()+d.getDuration()).toString());
 			if (dates.length>0) {estimatedCompDateField.setValue(dates[dates.length-1].getStart());
 			                     estimatedCompTimeField.setValue(dates[dates.length-1].getStart());}
 			else JOptionPane.showMessageDialog(null,"could not find a suitable date");
